Enum rusty_tcl::error::TclError

pub enum TclError {
    NullPointer,
    NulBytes(String),
    InvalidUtf8,
    InternalError(String),
}

Represents different classes of Tcl Errors.

Variants
NullPointer

Represents a pointer being NULL when it shouldn't have been. This usually happens in functions like [TclObj::new].

NulBytes(String)

Represents a string containing NUL bytes when it shouldn't have. This can happen in pretty much any function that takes a String.

InvalidUtf8

Represents a C string containing bytes that are invalid UTF-8.

InternalError(String)

Represents an error returned by Tcl itself.