Struct rusty_tcl::interp::TclInterp

pub struct TclInterp { /* fields omitted */ }

Interpreter struct that holds the Tcl interpreter itself.

Methods

impl TclInterp

pub fn new() -> Result<Self, TclError>

Creates a new interpreter.

Errors

This function returns an Err value if: 1. The pointer returned by Tcl_CreateInterp is NULL. 2. Initialization of the current interpreter via Tcl_Init fails.

pub fn make_safe(&mut self) -> Result<(), TclError>

Makes this interpreter safe.

Errors

This function returns an error when Tcl_MakeSafe fails.

Notes

As noted in the Tcl_MakeSafe man page, a "safe" interpreter only removes core potentially-unsafe functions. It's your responsibility to make sure any extensions you use are safe.

pub fn is_safe(&self) -> bool

Returns true if the current interpreter is safe.

pub fn eval_to_string(&mut self, code: &str) -> Result<&str, TclError>

Evaluates code, returning the interpreter's string result.

Errors

This function returns any errors that happen in the code as a Err(TclError::InternalError(..)), and it also returns any errors that happen when getting the interpreter's string result.

pub fn eval_to_obj(&mut self, code: &str) -> Result<TclObj, TclError>

Evaluates code, returning the interpreter's object result.

Errors

This function returns any errors that happen in the code as a Err(TclError::InternalError(..)), and it also returns any errors that happen when getting the interpreter's object result.

pub fn set_var(
&mut self,
name: impl Into<Vec<u8>>,
value: impl Into<Vec<u8>>
) -> Result<&CStr, TclError>

Sets a variable with the given name to the given value.

Errors

This function an error if either name or value contain NUL bytes, or if the pointer returned by Tcl_SetVar is NULL.

Notes

This returns the value that name was set to, which may differ from value due to tracing.